chore: convert uint32 to uint64

+++ dto/controller/inbound/extend_inbound.go
@@ -3,5 +3,5 @@ package inbound
 type ExtendInboundDto struct {
 	IdentifyInbounbdDto
 	End                  string `json:"end_time"`
-	ExtendedTrafficLimit uint32 `json:"extended_traffic_limit"`
+	ExtendedTrafficLimit uint64 `json:"extended_traffic_limit"`
 }
