+++ dto/service/inbound/extend_time.go
@@ -4,5 +4,5 @@ import "time"
 
 type ExtendInboundDto struct {
 	End                  time.Time
-	ExtendedTrafficLimit uint32
+	ExtendedTrafficLimit uint64
 }
